chore: move some shared helper to core
