+++ packages/core/src/server/devServer.ts
@@ -1,24 +1,21 @@
 import {
-  CreateDevMiddlewareReturns,
-  logger as defaultLogger,
-  StartDevServerOptions,
-  getAddressUrls,
-  printServerURLs,
   debug,
+  logger as defaultLogger,
   isFunction,
-  StartServerResult,
-  getDevOptions,
   ROOT_DIST_DIR,
-  formatRoutes,
+  getAddressUrls,
+  StartServerResult,
   getPublicPathFromCompiler,
-  RspackMultiCompiler,
-  RspackCompiler,
-  CompileMiddlewareAPI,
-  DevMiddlewaresConfig,
-  Routes,
-  DevServerAPIs,
+  type Routes,
+  type DevServerAPIs,
+  type RspackCompiler,
+  type RspackMultiCompiler,
+  type CompileMiddlewareAPI,
+  type DevMiddlewaresConfig,
+  type StartDevServerOptions,
+  type CreateDevMiddlewareReturns,
 } from '@rsbuild/shared';
-
+import { formatRoutes, getDevOptions, printServerURLs } from './helper';
 import connect from '@rsbuild/shared/connect';
 import { onBeforeRestartServer } from './restart';
 import type { Context } from '../types';
