+++ packages/core/tests/html.test.ts
@@ -1,4 +1,6 @@
 import { hasTitle } from '@src/rspack/HtmlBasicPlugin';
+import { getHTMLPathByEntry } from '@src/provider/core/initPlugins';
+import type { NormalizedConfig } from '@rsbuild/shared';
 
 test('should detect HTML title via "hasTitle" correctly', () => {
   expect(hasTitle()).toEqual(false);
@@ -18,4 +20,36 @@ test('should detect HTML title via "hasTitle" correctly', () => {
       '<html>\n<head>\n<title>\nPage Title\n</title>\n</head>\n<body></body>\n</html>',
     ),
   ).toEqual(true);
+});
+
+describe('getHTMLPathByEntry', () => {
+  it('should use distPath.html as the folder', async () => {
+    const htmlPath = getHTMLPathByEntry('main', {
+      output: {
+        distPath: {
+          html: 'my-html',
+        },
+      },
+      html: {
+        outputStructure: 'nested',
+      },
+    } as NormalizedConfig);
+
+    expect(htmlPath).toEqual('my-html/main/index.html');
+  });
+
+  it('should allow to disable html folder', async () => {
+    const htmlPath = getHTMLPathByEntry('main', {
+      output: {
+        distPath: {
+          html: 'html',
+        },
+      },
+      html: {
+        outputStructure: 'flat',
+      },
+    } as NormalizedConfig);
+
+    expect(htmlPath).toEqual('html/main.html');
+  });
 });
