+++ packages/core/tests/plugins/splitChunks.test.ts
@@ -1,5 +1,9 @@
 import { createStubRsbuild } from '@rsbuild/test-helper';
-import { pluginSplitChunks } from '@src/plugins/splitChunks';
+import {
+  pluginSplitChunks,
+  MODULE_PATH_REGEX,
+  getPackageNameFromModulePath,
+} from '@src/plugins/splitChunks';
 import type { RsbuildConfig } from '@rsbuild/shared';
 
 describe('plugin-split-chunks', () => {
@@ -148,5 +152,51 @@ describe('plugin-split-chunks', () => {
     const config = await rsbuild.unwrapConfig();
 
     expect(config).toMatchSnapshot();
+  });
+});
+
+describe('getPackageNameFromModulePath', () => {
+  it('should parse correct path fragment in npm/yarn', async () => {
+    let modulePath = '/path/to/node_modules/@scope/package-name/index.js';
+    let [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
+    expect(scope).toBe('@scope');
+    expect(name).toBe('package-name');
+
+    modulePath = '/path/to/node_modules/package-name/index.js';
+    [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
+    expect(scope).toBe(undefined);
+    expect(name).toBe('package-name');
+  });
+
+  it('should parse correct path fragment in pnpm', async () => {
+    let modulePath = '/path/to/node_modules/.pnpm/@scope/package-name/index.js';
+    let [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
+    expect(scope).toBe('@scope');
+    expect(name).toBe('package-name');
+
+    modulePath = '/path/to/node_modules/.pnpm/package-name/index.js';
+    [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
+    expect(scope).toBe(undefined);
+    expect(name).toBe('package-name');
+  });
+
+  it('should return correct package name in npm/yarn', () => {
+    let modulePath = '/path/to/node_modules/@scope/package-name/index.js';
+    expect(getPackageNameFromModulePath(modulePath)).toBe(
+      'npm.scope.package-name',
+    );
+
+    modulePath = '/path/to/node_modules/package-name/index.js';
+    expect(getPackageNameFromModulePath(modulePath)).toBe('npm.package-name');
+  });
+
+  it('should return correct package name in pnpm', () => {
+    let modulePath = '/path/to/node_modules/.pnpm/@scope/package-name/index.js';
+    expect(getPackageNameFromModulePath(modulePath)).toBe(
+      'npm.scope.package-name',
+    );
+
+    modulePath = '/path/to/node_modules/.pnpm/package-name/index.js';
+    expect(getPackageNameFromModulePath(modulePath)).toBe('npm.package-name');
   });
 });
