--- packages/shared/tests/fs.test.ts
@@ -1,84 +1,0 @@
-import {
-  getPackageNameFromModulePath,
-  MODULE_PATH_REGEX,
-  getHTMLPathByEntry,
-  NormalizedOutputConfig,
-} from '../src';
-
-describe('getPackageNameFromModulePath', () => {
-  it('should parse correct path fragment in npm/yarn', async () => {
-    let modulePath = '/path/to/node_modules/@scope/package-name/index.js';
-    let [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
-    expect(scope).toBe('@scope');
-    expect(name).toBe('package-name');
-
-    modulePath = '/path/to/node_modules/package-name/index.js';
-    [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
-    expect(scope).toBe(undefined);
-    expect(name).toBe('package-name');
-  });
-
-  it('should parse correct path fragment in pnpm', async () => {
-    let modulePath = '/path/to/node_modules/.pnpm/@scope/package-name/index.js';
-    let [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
-    expect(scope).toBe('@scope');
-    expect(name).toBe('package-name');
-
-    modulePath = '/path/to/node_modules/.pnpm/package-name/index.js';
-    [, , scope, name] = modulePath.match(MODULE_PATH_REGEX)!;
-    expect(scope).toBe(undefined);
-    expect(name).toBe('package-name');
-  });
-
-  it('should return correct package name in npm/yarn', () => {
-    let modulePath = '/path/to/node_modules/@scope/package-name/index.js';
-    expect(getPackageNameFromModulePath(modulePath)).toBe(
-      'npm.scope.package-name',
-    );
-
-    modulePath = '/path/to/node_modules/package-name/index.js';
-    expect(getPackageNameFromModulePath(modulePath)).toBe('npm.package-name');
-  });
-
-  it('should return correct package name in pnpm', () => {
-    let modulePath = '/path/to/node_modules/.pnpm/@scope/package-name/index.js';
-    expect(getPackageNameFromModulePath(modulePath)).toBe(
-      'npm.scope.package-name',
-    );
-
-    modulePath = '/path/to/node_modules/.pnpm/package-name/index.js';
-    expect(getPackageNameFromModulePath(modulePath)).toBe('npm.package-name');
-  });
-});
-
-describe('getHTMLPathByEntry', () => {
-  it('should use distPath.html as the folder', async () => {
-    const htmlPath = getHTMLPathByEntry('main', {
-      output: {
-        distPath: {
-          html: 'my-html',
-        },
-      } as NormalizedOutputConfig,
-      html: {
-        outputStructure: 'nested',
-      },
-    });
-
-    expect(htmlPath).toEqual('my-html/main/index.html');
-  });
-
-  it('should allow to disable html folder', async () => {
-    const htmlPath = getHTMLPathByEntry('main', {
-      output: {
-        distPath: {
-          html: 'html',
-        },
-      } as NormalizedOutputConfig,
-      html: {
-        outputStructure: 'flat',
-      },
-    });
-
-    expect(htmlPath).toEqual('html/main.html');
-  });
-});
